[auth] fix: move class AllIconData to models
