+++ auth/lib/models/all_icon_data.dart
@@ -1,0 +1,15 @@
+enum IconType { simpleIcon, customIcon }
+
+class AllIconData {
+  final String title;
+  final IconType type;
+  final String? color;
+  final String? slug;
+
+  AllIconData({
+    required this.title,
+    required this.type,
+    required this.color,
+    this.slug,
+  });
+}
